Add: new entities for teachhub

+++ backend/src/main/java/teachhub/com/TeachHub/model/roles/RolesRepository.java
@@ -2,5 +2,5 @@ package teachhub.com.TeachHub.model.roles;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RolesRepository extends JpaRepository<Roles, Integer> {
+public interface RolesRepository extends JpaRepository<Roles, Long> {
 }
